Config database link to test database
This is the config file linking to the database. This config is
important and should not be changed in test branch. An option is
adding it to .gitignore to avoid it been changed by merge. Also,
this file in production branch should also be taken care to avoid
the misuse of database.